agentic/requirements.txt
# Core dependencies
crewai>=0.28.0
langchain>=0.0.267
langchain-openai>=0.0.2

# PowerPoint manipulation
python-pptx>=0.6.21

# Frontend
streamlit>=1.24.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.3
numpy>=1.24.3

# API integrations
openai>=0.27.8
tavily-python>=0.2.6
